+++ Extract/cpz/PB3B.h
@@ -15,15 +15,15 @@ public:
 private:
 	void Decrypt(u8* target, size_t size);
 
-	bool Decode1(CArcFile* archive, const u8* src, size_t src_size, long width, long height, u16 bpp);
-	bool Decode3(CArcFile* archive, const u8* src, size_t src_size, long width, long height, u16 bpp);
-	bool Decode4(CArcFile* archive, const u8* src, size_t src_size, long width, long height, u16 bpp);
-	bool Decode5(CArcFile* archive, const u8* src, size_t src_size, long width, long height, u16 bpp);
-	bool Decode6(CArcFile* archive, const u8* src, size_t src_size, long width, long height, u16 bpp, PB3B_DECRYPT decryption_function);
+	bool Decode1(CArcFile* archive, const u8* src, size_t src_size, s32 width, s32 height, u16 bpp);
+	bool Decode3(CArcFile* archive, const u8* src, size_t src_size, s32 width, s32 height, u16 bpp);
+	bool Decode4(CArcFile* archive, const u8* src, size_t src_size, s32 width, s32 height, u16 bpp);
+	bool Decode5(CArcFile* archive, const u8* src, size_t src_size, s32 width, s32 height, u16 bpp);
+	bool Decode6(CArcFile* archive, const u8* src, size_t src_size, s32 width, s32 height, u16 bpp, PB3B_DECRYPT decryption_function);
 
-	bool Decomp1(u8* dst, size_t dst_size, const u8* src, size_t src_size, long width, long height, u16 bpp);
-	bool Decomp3(u8* dst, size_t dst_size, const u8* src, size_t src_size, long width, long height, u16 bpp);
-	bool Decomp4(u8* dst, size_t dst_size, const u8* src, size_t src_size, long width, long height, u16 bpp);
-	bool Decomp5(u8* dst, size_t dst_size, const u8* src, size_t src_size, long width, long height, u16 bpp);
-	bool Decomp6(u8* dst, size_t dst_size, const u8* src, size_t src_size, long width, long height, u16 bpp);
+	bool Decomp1(u8* dst, size_t dst_size, const u8* src, size_t src_size, s32 width, s32 height, u16 bpp);
+	bool Decomp3(u8* dst, size_t dst_size, const u8* src, size_t src_size, s32 width, s32 height, u16 bpp);
+	bool Decomp4(u8* dst, size_t dst_size, const u8* src, size_t src_size, s32 width, s32 height, u16 bpp);
+	bool Decomp5(u8* dst, size_t dst_size, const u8* src, size_t src_size, s32 width, s32 height, u16 bpp);
+	bool Decomp6(u8* dst, size_t dst_size, const u8* src, size_t src_size, s32 width, s32 height, u16 bpp);
 };
